reafctor: querybuilder
- update() and save() will return the model
- delete, destroy will return affected rows count
